requirements.txt
easyocr>=1.7.0
openpyxl>=3.1.0
ttkbootstrap>=1.10.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
